feature: Join popBackStack 추가 및 Navigation 추가
